requirements.txt
python-telegram-bot==21.3
python-dotenv==1.0.1

